Developer Name: Kristian Synopsis: Know errors: The "Slutbilled" is always white. When attaching or adding a slutbilled, there is always a highlighted item. lastUsed list is reset after each close down.
Fixed errors:
Attachment would be set to null when you changed it to a timetimer.
DoneScreenPicture would be set to null if you changed to timer type or tried to save the subprofile.
Attachment is now loaded and saved correctly from the database, no matter if it is Timer, SingleImg, or SplitImg.
Slutbilled is now loaded and saved correctly from the database, no matter if it is SingleImg, or SplitImg.

Changes:
Changed the certificate in our owned dummy data which is in the Guardian class, Mette Als got the same certificate as "Person A".
Fixed the DoneScreen (Slutbilled) button so it changes according to the subprofile.
Removed all static properties and methods in Guardian, only _instance is static to make sure it is singleton.
Charged all static references in wombat project.

Notes:
lastUsed is sometimes crashing, it is therefore been commented, so it do not affect the application.